changed cashback field type

+++ src/lib/validations/cashback.ts
@@ -3,7 +3,7 @@ import * as z from 'zod'
 export const cashbackSchema = z.object({
   retailerId: z.string({ required_error: 'Selecione um varejista' }),
   provider: z.string().min(1, 'Campo obrigatório'),
-  value: z.coerce.number({ required_error: 'Campo obrigatório' }).int(),
+  value: z.coerce.number({ required_error: 'Campo obrigatório' }),
   url: z
     .string()
     .min(1, 'Campo obrigatório')
